+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- `Message.Builder#setBody(byte[])` to avoid `ByteBuffer.wrap` in the client code. 
+- `Message.Builder#setBody(byte[])` to avoid `ByteBuffer.wrap` in the client code.
+- `MapIndex.isEmpty()` method to check if MapIndex is empty.
 
 ### Changed
 - `Transaction#execute` can throw `TransactionExecutionException` to roll back